large refactor and more checkmate tests
This changes in-check checking to happen interior to is_movement_valid instead
of in parse_algebraic; this makes a bunch more sense (because is_movement_valid
is the real API for movement checking, not parse_algebraic). This allows for
more robust in-check checking too, which in turn allows for better checkmate
detection.

Checkmate detection still false-positives when a piece could block the attack of
the threatening piece.