fix find_all_with_access with no results with glibc
OpenBSD's malloc returns a "valid" pointer to a structure of zero length
when requesting a realloc; the given pointer has these properties:

    Allocation of a zero size object returns a pointer to a zero
    size object.  This zero size object is access protected, so any
    access to it will generate an exception (SIGSEGV).  Many
    zero-sized objects can be placed consecutively in shared
    protected pages.  The minimum size of the protection on each
    object is suitably aligned and sized as previously stated, but
    the protection may extend further depending on where in a
    protected zone the object lands.

This is really cool but unfortunately not the case on other platforms,
like, say, Linux, where realloc(x, 0) returns NULL. Our NULL checking
for when realloc fails itself then fails, even though the NULL result is
the intended result.

This change modifies find_all_with_access to expect a NULL return result
when we're trying to "create" a list of zero length, preventing the
double-free that was previously occurring.